requirements.txt
SQLAlchemy==1.3.9
requests==2.22.0
pycep_correios==4.0.1
python-dotenv==0.12.0
lxml==4.5.0
